+++ railties/test/generators/scaffold_generator_test.rb
@@ -430,7 +430,7 @@ class ScaffoldGeneratorTest < Rails::Generators::TestCase
 
     assert_file "app/controllers/accounts_controller.rb" do |content|
       assert_instance_method :account_params, content do |m|
-        assert_match(/expect\(account: \[:name, :currency_id, :user_id\]\)/, m)
+        assert_match(/expect\(account: \[ :name, :currency_id, :user_id \]\)/, m)
       end
     end
 
@@ -461,7 +461,7 @@ class ScaffoldGeneratorTest < Rails::Generators::TestCase
 
     assert_file "app/controllers/messages_controller.rb" do |content|
       assert_instance_method :message_params, content do |m|
-        assert_match(/expect\(message: \[:video, photos: \[\], images: \[\]\]\)/, m)
+        assert_match(/expect\(message: \[ :video, photos: \[\], images: \[\] \]\)/, m)
       end
     end
 
@@ -488,7 +488,7 @@ class ScaffoldGeneratorTest < Rails::Generators::TestCase
 
     assert_file "app/controllers/messages_controller.rb" do |content|
       assert_instance_method :message_params, content do |m|
-        assert_match(/expect\(message: \[:content\]\)/, m)
+        assert_match(/expect\(message: \[ :content \]\)/, m)
       end
     end
 
@@ -536,7 +536,7 @@ class ScaffoldGeneratorTest < Rails::Generators::TestCase
 
     assert_file "app/controllers/users_controller.rb" do |content|
       assert_instance_method :user_params, content do |m|
-        assert_match(/expect\(user: \[:name, :password, :password_confirmation\]\)/, m)
+        assert_match(/expect\(user: \[ :name, :password, :password_confirmation \]\)/, m)
       end
     end